Stop using audit in tests
