+++ src/ServiceControl.AcceptanceTests/MessageFailures/Is_System_Message_Tests.cs
@@ -131,7 +131,7 @@
         {
             public ServerEndpoint()
             {
-                EndpointSetup<DefaultServerWithAudit>();
+                EndpointSetup<DefaultServerWithoutAudit>();
             }
 
             class Foo : DispatchRawMessages<SystemMessageTestContext>
