+++ src/ServiceControl.AcceptanceTests/MessageFailures/When_a_retry_for_a_failed_message_fails.cs
@@ -100,7 +100,7 @@
         {
             public FailureEndpoint()
             {
-                EndpointSetup<DefaultServerWithAudit>(c => { c.NoRetries(); });
+                EndpointSetup<DefaultServerWithoutAudit>(c => { c.NoRetries(); });
             }
 
             public class MyMessageHandler : IHandleMessages<MyMessage>
