+++ src/ServiceControl.AcceptanceTests/MessageFailures/When_a_retry_for_a_failed_message_is_successful.cs
@@ -202,7 +202,7 @@
         {
             public FailureEndpoint()
             {
-                EndpointSetup<DefaultServerWithAudit>(c => { c.NoRetries(); });
+                EndpointSetup<DefaultServerWithoutAudit>(c => { c.NoRetries(); });
             }
 
             public class MyMessageHandler : IHandleMessages<MyMessage>
