+++ src/ServiceControl.AcceptanceTests/MessageFailures/When_processing_message_with_missing_metadata_failed.cs
@@ -77,7 +77,7 @@
         {
             public Failing()
             {
-                EndpointSetup<DefaultServerWithAudit>(c => { c.Recoverability().Delayed(x => x.NumberOfRetries(0)); });
+                EndpointSetup<DefaultServerWithoutAudit>(c => { c.Recoverability().Delayed(x => x.NumberOfRetries(0)); });
             }
 
             class SendFailedMessage : DispatchRawMessages<MyContext>
