+++ src/ServiceControl.AcceptanceTests/SagaAudit/When_a_message_emitted_by_a_saga_is_audited.cs
@@ -35,7 +35,7 @@
         {
             public SagaEndpoint()
             {
-                EndpointSetup<DefaultServerWithAudit>();
+                EndpointSetup<DefaultServerWithoutAudit>();
             }
 
             public class MySaga : Saga<MySagaData>, IAmStartedByMessages<MessageInitiatingSaga>
