+++ src/ServiceControl.AcceptanceTests/SagaAudit/When_a_message_hitting_a_saga_is_audited.cs
@@ -44,7 +44,7 @@
         {
             public SagaEndpoint()
             {
-                EndpointSetup<DefaultServerWithAudit>(c => c.AuditSagaStateChanges(Settings.DEFAULT_SERVICE_NAME));
+                EndpointSetup<DefaultServerWithoutAudit>(c => c.AuditSagaStateChanges(Settings.DEFAULT_SERVICE_NAME));
             }
 
             public class MySaga : Saga<MySagaData>, IAmStartedByMessages<MessageInitiatingSaga>
