+++ src/ServiceControl.AcceptanceTests/SagaAudit/When_a_message_hitting_a_saga_is_not_a_start_message.cs
@@ -24,7 +24,7 @@
         {
             public SagaEndpoint()
             {
-                EndpointSetup<DefaultServerWithAudit>();
+                EndpointSetup<DefaultServerWithoutAudit>();
             }
 
             public class MySaga : Saga<MySagaData>, IAmStartedByMessages<MessageInitiatingSaga>,
