+++ src/ServiceControl.AcceptanceTests/SagaAudit/When_a_saga_instance_is_being_created.cs
@@ -42,7 +42,7 @@
         {
             public SagaEndpoint()
             {
-                EndpointSetup<DefaultServerWithAudit>(
+                EndpointSetup<DefaultServerWithoutAudit>(
                     c => c.AuditSagaStateChanges(Settings.DEFAULT_SERVICE_NAME));
             }
         }
