+++ src/ServiceControl.AcceptanceTests/SagaAudit/When_multiple_messages_are_emitted_by_a_saga.cs
@@ -60,7 +60,7 @@
         {
             public SagaEndpoint()
             {
-                EndpointSetup<DefaultServerWithAudit>(c =>
+                EndpointSetup<DefaultServerWithoutAudit>(c =>
                 {
                     // NOTE: The default template disables this feature but that means the event will not be subscribed to or published
                     c.EnableFeature<AutoSubscribe>();
