+++ src/ServiceControl.AcceptanceTests/SagaAudit/When_requesting_timeout_from_a_saga.cs
@@ -40,7 +40,7 @@
         {
             public SagaEndpoint()
             {
-                EndpointSetup<DefaultServerWithAudit>(c => c.AuditSagaStateChanges(Settings.DEFAULT_SERVICE_NAME));
+                EndpointSetup<DefaultServerWithoutAudit>(c => c.AuditSagaStateChanges(Settings.DEFAULT_SERVICE_NAME));
             }
         }
